requirements.txt
aiohttp==4.0.0a1
aioredis==1.3.1
appdirs==1.4.3
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
entrypoints==0.3
fastapi==0.52.0
flake8==3.7.9
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
idna==2.9
mccabe==0.6.1
multidict==4.7.5
mypy==0.761
mypy-extensions==0.4.3
pathspec==0.7.0
pycodestyle==2.5.0
pydantic==1.4
pyflakes==2.1.1
python-dateutil==2.8.1
regex==2020.2.20
requests==2.23.0
six==1.14.0
starlette==0.13.2
toml==0.10.0
typed-ast==1.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
yarl==1.4.2
